+++ tests/integration/concrete/batch/batch.t.sol
@@ -227,8 +227,8 @@ contract Batch_Integration_Concrete_Test is Integration_Test {
         calls[0] = abi.encodeCall(flow.pause, (defaultStreamIds[0]));
         calls[1] = abi.encodeCall(flow.pause, (defaultStreamIds[1]));
 
-        uint128 previousTotalDebt0 = flow.totalDebtOf(defaultStreamId);
-        uint128 previousTotalDebt1 = flow.totalDebtOf(defaultStreamIds[1]);
+        uint256 previousTotalDebt0 = flow.totalDebtOf(defaultStreamId);
+        uint256 previousTotalDebt1 = flow.totalDebtOf(defaultStreamIds[1]);
 
         // It should emit 2 {PauseFlowStream}, 2 {MetadataUpdate} events.
