+++ tests/integration/concrete/pause/pause.t.sol
@@ -66,7 +66,7 @@ contract Pause_Integration_Concrete_Test is Integration_Test {
     }
 
     function test_Pause() internal {
-        uint128 initialTotalDebt = flow.totalDebtOf(defaultStreamId);
+        uint256 initialTotalDebt = flow.totalDebtOf(defaultStreamId);
 
         // It should emit 1 {PauseFlowStream}, 1 {MetadataUpdate} events.
         vm.expectEmit({ emitter: address(flow) });
@@ -90,7 +90,7 @@ contract Pause_Integration_Concrete_Test is Integration_Test {
         assertEq(actualRatePerSecond, 0, "rate per second");
 
         // It should update the snapshot debt.
-        uint128 actualSnapshotDebt = flow.getSnapshotDebt(defaultStreamId);
+        uint256 actualSnapshotDebt = flow.getSnapshotDebt(defaultStreamId);
         assertEq(actualSnapshotDebt, initialTotalDebt, "snapshot debt");
     }
 }
